+++ backend/Application/Extensions/StartupExtensions.cs
@@ -1,9 +1,17 @@
-using System.Runtime.CompilerServices;
+using System.IO;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
 using Application.Configurations;
 using Application.NodeApi;
 using Concordium.Sdk.Client;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Prometheus;
 
 namespace Application.Extensions;
 
@@ -16,7 +24,22 @@ internal static class StartupExtensions
         var uri = new Uri(grpcNodeClientSettings.Address);
         services.AddSingleton(new ConcordiumClient(uri, concordiumClientOptions));
     }
-    
+
+    internal static void AddDefaultHealthChecks(this IServiceCollection services)
+    {
+        services.AddHealthChecks()
+            .AddCheck("live", () => HealthCheckResult.Healthy("Application is running"))
+            .ForwardToPrometheus();
+    }
+
+    internal static void AddHealthChecks(this IApplicationBuilder app)
+    {
+        app.UseHealthChecks("/system/health", new HealthCheckOptions
+        {
+            ResponseWriter = WriteResponse
+        });
+    }
+
     internal static IServiceCollection AddFeatureFlagOptions(this IServiceCollection collection, IConfiguration configuration, ILogger logger)
     {
         var featureFlags = configuration.GetSection("FeatureFlags").Get<FeatureFlagOptions>();
@@ -25,5 +48,51 @@ internal static class StartupExtensions
         logger.Information("Feature flag [{name}]: {value}", nameof(featureFlags.ConcordiumNodeImportEnabled), featureFlags.ConcordiumNodeImportEnabled);
         logger.Information("Feature flag [{name}]: {value}", nameof(featureFlags.ConcordiumNodeImportValidationEnabled), featureFlags.ConcordiumNodeImportValidationEnabled);
         return collection;
+    }    
+    
+    /// <summary>
+    /// Taken from https://learn.microsoft.com/en-us/aspnet/core/host-and-deploy/health-checks?view=aspnetcore-7.0#customize-output
+    /// such that health state is exported.
+    /// </summary>
+    private static Task WriteResponse(HttpContext context, HealthReport healthReport)
+    {
+        context.Response.ContentType = "application/json; charset=utf-8";
+
+        var options = new JsonWriterOptions { Indented = true };
+
+        using var memoryStream = new MemoryStream();
+        using (var jsonWriter = new Utf8JsonWriter(memoryStream, options))
+        {
+            jsonWriter.WriteStartObject();
+            jsonWriter.WriteString("status", healthReport.Status.ToString());
+            jsonWriter.WriteStartObject("results");
+
+            foreach (var healthReportEntry in healthReport.Entries)
+            {
+                jsonWriter.WriteStartObject(healthReportEntry.Key);
+                jsonWriter.WriteString("status",
+                    healthReportEntry.Value.Status.ToString());
+                jsonWriter.WriteString("description",
+                    healthReportEntry.Value.Description);
+                jsonWriter.WriteStartObject("data");
+
+                foreach (var item in healthReportEntry.Value.Data)
+                {
+                    jsonWriter.WritePropertyName(item.Key);
+
+                    JsonSerializer.Serialize(jsonWriter, item.Value,
+                        item.Value?.GetType() ?? typeof(object));
+                }
+
+                jsonWriter.WriteEndObject();
+                jsonWriter.WriteEndObject();
+            }
+
+            jsonWriter.WriteEndObject();
+            jsonWriter.WriteEndObject();
+        }
+
+        return context.Response.WriteAsync(
+            Encoding.UTF8.GetString(memoryStream.ToArray()));
     }
 }