+++ backend/Tests/Api/GraphQL/Import/ImportStateControllerTest.cs
@@ -5,7 +5,6 @@ using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using Tests.TestUtilities;
 using Tests.TestUtilities.Stubs;
-using Xunit.Abstractions;
 
 namespace Tests.Api.GraphQL.Import;