+++ backend/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 - Added feature which fetched Smart Contract data from node and store as events.
+- Added Prometheus metrics which give durations of imports, count of relevant processed transaction events and retry counts.
+- Added health checks, where application goes into degraded state if some job stops running. These health checks are now making endpoint `rest/health` redundant and it is removed.
 
 ## 1.6.3
 - Added Account Balance to Account Statement export file.